fix(read-guard): track bash writes too, consistent with the Write tool (refs #168)
Per review: a bash write means the agent authored the file, so — exactly like the
Write tool — it should NOT block a follow-up edit. Mirror the existing read-guard
pattern for both directions instead of dropping writes:

  bash READS  (cat/head/tail/sed -n) → recordRead with the exact line range shown
  bash WRITES (>, >>, N>, tee, sed -i, cp/mv dest, touch) → noteCreatedFile at
              tool_call + recordWritten at tool_result, the same two-phase flow the
              Write tool uses (adds the authoritative writtenThisSession entry and
              injects the creation read).

Still NOT registered: grep (scattered matches), find/ls (names only), and bare
path mentions — no contiguous content is shown.

- New clients/bash-file-access.ts (moved out of index.ts to avoid a circular
  import from runtime-tool-result.ts): extractReadPathsFromCommand (ReadSpan[]) +
  extractWrittenPathsFromCommand (string[]). ReDoS-safe, verb-based per shell
  segment.
- index.ts tool_call: recordRead for reads, noteCreatedFile for writes.
- runtime-tool-result.ts: recordWritten for bash write targets.
- Tests rewritten (29): range assertions for reads, write-target assertions for
  every write form, and negatives proving grep/find/ls/bare mentions register
  neither a read nor a write.
